# Speedup transproc `map_array!`