@splattael 2015-01-08
After this second commit one can see even more speedup:

```
$ ruby -Ilib foo.rb 
Calculating -------------------------------------
        map_array! 1    24.725k i/100ms
       map_array! 10    11.440k i/100ms
      map_array! 100     1.806k i/100ms
     map_array! 1000   194.000  i/100ms
-------------------------------------------------
        map_array! 1    660.707k (± 2.7%) i/s -      3.313M
       map_array! 10    161.511k (± 2.7%) i/s -    812.240k
      map_array! 100     18.791k (± 1.9%) i/s -     95.718k
     map_array! 1000      1.830k (±17.0%) i/s -      8.924k
```