@solnic 2015-01-08
This is so sweet Peter.

BTW you can use `x.compare!` with bechmark/ips and compare new version of a function against the old one (just keep the old one under a different name for benchmark purposes).